refactor: reorganize project structure by moving collections and plugins to modules; update import paths accordingly
