+++ src/app/(frontend)/search/page.tsx
@@ -4,8 +4,7 @@ import { CollectionArchive } from '@/components/CollectionArchive'
 import configPromise from '@payload-config'
 import { getPayload } from 'payload'
 import React from 'react'
-import { Post } from '@/payload-types'
-import { Search } from '@/search/Component'
+import { Search } from '@/modules/Settings/Search/Component'
 import PageClient from './page.client'
 import { CardPostData } from '@/components/Card'
