+++ src/modules/Content/Pages/index.ts
@@ -1,16 +1,16 @@
 import type { CollectionConfig } from 'payload'
 
-import { authenticated } from '../../access/authenticated'
-import { authenticatedOrPublished } from '../../access/authenticatedOrPublished'
-import { Archive } from '../../blocks/ArchiveBlock/config'
-import { CallToAction } from '../../blocks/CallToAction/config'
-import { Content } from '../../blocks/Content/config'
-import { FormBlock } from '../../blocks/Form/config'
-import { MediaBlock } from '../../blocks/MediaBlock/config'
+import { authenticated } from '../../../access/authenticated'
+import { authenticatedOrPublished } from '../../../access/authenticatedOrPublished'
+import { Archive } from '../../../blocks/ArchiveBlock/config'
+import { CallToAction } from '../../../blocks/CallToAction/config'
+import { Content } from '../../../blocks/Content/config'
+import { FormBlock } from '../../../blocks/Form/config'
+import { MediaBlock } from '../../../blocks/MediaBlock/config'
 import { hero } from '@/heros/config'
 import { slugField } from '@/fields/slug'
-import { populatePublishedAt } from '../../hooks/populatePublishedAt'
-import { generatePreviewPath } from '../../utilities/generatePreviewPath'
+import { populatePublishedAt } from '../../../hooks/populatePublishedAt'
+import { generatePreviewPath } from '../../../utilities/generatePreviewPath'
 import { revalidatePage } from './hooks/revalidatePage'
 
 import {
