+++ src/modules/Content/Posts/hooks/revalidatePost.ts
@@ -2,7 +2,7 @@ import type { CollectionAfterChangeHook } from 'payload'
 
 import { revalidatePath } from 'next/cache'
 
-import type { Post } from '../../../payload-types'
+import type { Post } from '../../../../payload-types'
 
 export const revalidatePost: CollectionAfterChangeHook<Post> = ({
   doc,
