+++ src/modules/Content/Posts/index.ts
@@ -9,12 +9,12 @@ import {
   lexicalEditor,
 } from '@payloadcms/richtext-lexical'
 
-import { authenticated } from '../../access/authenticated'
-import { authenticatedOrPublished } from '../../access/authenticatedOrPublished'
-import { Banner } from '../../blocks/Banner/config'
-import { Code } from '../../blocks/Code/config'
-import { MediaBlock } from '../../blocks/MediaBlock/config'
-import { generatePreviewPath } from '../../utilities/generatePreviewPath'
+import { authenticated } from '../../../access/authenticated'
+import { authenticatedOrPublished } from '../../../access/authenticatedOrPublished'
+import { Banner } from '../../../blocks/Banner/config'
+import { Code } from '../../../blocks/Code/config'
+import { MediaBlock } from '../../../blocks/MediaBlock/config'
+import { generatePreviewPath } from '../../../utilities/generatePreviewPath'
 import { populateAuthors } from './hooks/populateAuthors'
 import { revalidatePost } from './hooks/revalidatePost'
