+++ src/modules/Navigation/Footer/config.ts
