+++ src/modules/Settings/Redirects/config.ts
