+++ src/plugins/index.ts
@@ -1,11 +1,11 @@
-import { payloadCloudPlugin } from '@payloadcms/payload-cloud'
 import { Plugin } from 'payload'
-import { s3StoragePlugin } from './s3Storage'
-import { redirectsPluginConfig } from './redirects'
-import { nestedDocsPluginConfig } from './nestedDocs'
-import { seoPluginConfig } from './seo'
-import { formBuilderPluginConfig } from './formBuilder'
-import { searchPluginConfig } from './search'
+import { payloadCloudPlugin } from '@payloadcms/payload-cloud'
+import { s3StoragePlugin } from '../modules/Upload/s3Storage'
+import { nestedDocsPluginConfig } from '../modules/Content/nestedDocs'
+import { seoPluginConfig } from '../modules/Content/seo'
+import { formBuilderPluginConfig } from '../modules/Settings/FormBuilder/config'
+import { redirectsPluginConfig } from '../modules/Settings/Redirects/config'
+import { searchPluginConfig } from '../modules/Settings/Search/config'
 
 export const plugins: Plugin[] = [
   s3StoragePlugin,
